requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.5.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pypdfium2==4.30.0
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.43
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
